// tcc_pv_front/src/app/models/Residente.ts
export class Residente
{
  id: string = '';
  Dataregistro: string = '';
      Nome: string = '';
      DataNasc: string = '';
      genero: string ='';
      aposentado: boolean= true ;
      endereco: string= '';
      emprestimonome: boolean= true;
      emprestimonomepessoa: string= '';
      grauldependencia: string= '';
      deficiencia: boolean= true;
      deficiencianome: string= '';
      doenca: boolean= true;
      doencanome: string= '';
      remediocontrol: boolean=true;
      remediocontrolnome: string= '';
      motivoprocura: string = '';
      nomeprocura: string= '';
      grauparente: string= '';
      enderecoparente: string= '';
      telefone: string= '';
}